+++ blocks/battery.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 )
 
+// Battery returns the current percentage remaining.
 func Battery() (string, error) {
 	cmd := "cat /sys/class/power_supply/BAT0/capacity"
 	runCmd, err := exec.Command("sh", "-c", cmd).Output()
@@ -14,6 +15,6 @@ func Battery() (string, error) {
 	}
 
 	perc := "%"
-	level := fmt.Sprintf("%s%s |", strings.TrimSpace(string(runCmd)), perc)
+	level := fmt.Sprintf("%s%s", strings.TrimSpace(string(runCmd)), perc)
 	return level, nil
 }