+++ blocks/volume.go
@@ -6,6 +6,7 @@ import (
 	"strings"
 )
 
+// VolumeText sends back the current volume percent.
 func VolumeText() (string, error) {
 	volCmd := "amixer -D pulse sget Master | awk '/Front Right:/ {print $5}' | grep -o '[0-9].'"
 	runVol, err := exec.Command("sh", "-c", volCmd).Output()
@@ -14,5 +15,5 @@ func VolumeText() (string, error) {
 	}
 
 	percent := "%"
-	return fmt.Sprintf("%s%s |", strings.TrimSpace(string(runVol)), percent), nil
+	return fmt.Sprintf("%s%s", strings.TrimSpace(string(runVol)), percent), nil
 }